+++ lib/ui/core/ui/widgets/dialog/edit_inventory_item_dialog.dart
@@ -70,9 +70,9 @@ class EditInventoryItemDialog extends HookConsumerWidget {
         purchaseDate.value,
         null,
         null,
-        int.tryParse(itemHeightController.text),
-        int.tryParse(itemWidthController.text),
-        int.tryParse(itemDepthController.text),
+        double.tryParse(itemHeightController.text),
+        double.tryParse(itemWidthController.text),
+        double.tryParse(itemDepthController.text),
         null,
         addToBooth.value ? 1 : 0,
